Traffic Light prototype
The prototype of the traffic light Tanda and Jie did — needs updating
